Move Web/Express functionality to web directory
