+++ test/web/api/index.test.js
@@ -6,18 +6,20 @@
 const test = require('ava');
 const supertest = require('supertest');
 require('chai').should();
-const app = require('../../../app');
+const app = require('../../../web/blinkWeb');
 
 /**
- * Test /api/v1.
+ * Test /api.
  */
-test('GET /api/v1 should list available endpoints', async () => {
-  const res = await supertest(app).get('/api/v1');
+test('GET /api should respond with JSON list of API versions', async () => {
+  const res = await supertest(app).get('/api');
   res.status.should.be.equal(200);
 
   // Check response to be json
   res.header.should.have.property('content-type');
   res.header['content-type'].should.match(/json/);
 
-  // No endpoints to test now.
+  // Check response to include /api/v1
+  res.body.should.have.property('v1');
+  res.body.v1.should.match(/\/api\/v1$/);
 });
