+++ test/web/blinkWeb.test.js
@@ -6,13 +6,13 @@
 const test = require('ava');
 const supertest = require('supertest');
 require('chai').should();
-const app = require('../app');
+const blinkWeb = require('../../web/blinkWeb');
 
 /**
  * Test root.
  */
 test('GET / should be polite', async () => {
-  const res = await supertest(app).get('/');
+  const res = await supertest(blinkWeb).get('/');
   res.status.should.be.equal(200);
   res.text.should.be.equal('Hi, I\'m Blink!');
 });
